+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# TBA
+
+- Add `with_relation_name` to `have_many` matcher
+
 # v0.2.0
 
 - Add `have_one` matcher